+++ src/base/b-list/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.?? (2023-??-??)
+
+#### :rocket: New Feature
+
+* Added new method `getHref`
+
 ## v3.43.0 (2023-03-23)
 
 #### :bug: Bug Fix